requirements.txt
altair==5.3.0
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imbalanced-learn==0.12.2
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
plotly-express==0.4.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.4.1.post1
scipy==1.13.1
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
streamlit==1.35.0
streamlit-authenticator==0.3.2
streamlit-lottie==0.0.5
tenacity==8.4.1
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
